fix: fix return to update and delete product services

+++ back-end/src/services/Products.service.js
@@ -21,12 +21,13 @@ const ProductsServices = {
     if (price < 0.5) throw new CodeError('Price must be greater than 0.5', 400);
     const exists = await model.getOne(id);
     if (!exists) throw new CodeError('Product not found', 404);
-    const product = await model.updatePrice({ id, price });
-    return product;
+    await model.updatePrice({ id, price });
+    return 'Product updated';
   },
   delete: async (id) => {
-    const product = await model.delete(id);
-    return product;
+    const exists = await model.getOne(id);
+    if (!exists) throw new CodeError('Product not found', 404);
+    await model.delete(id);
   },
 };
